Fix filenames in Project 10
