+++ projects/10/typescript-src/src/CompilationEngine.ts
@@ -20,17 +20,6 @@ export class CompilationEngine {
     this.results.push(`${this.indentation.spaces()}${value}`);
   }
 
-  // FIXME: このメソッドを使わなくて済むように直す
-  private popResults(): string {
-    const last = this.results.pop();
-    const regexp = /<.*/g;
-    const match = last?.match(regexp);
-    if (!match) {
-      return "";
-    }
-    return match[0];
-  }
-
   /**
    * 開始タグを挿入
    * @param tag
@@ -135,11 +124,11 @@ export class CompilationEngine {
     this.pushResults(`<${type}> ${this.tokenizer.keyWord()} </${type}>`);
     while (
       this.tokenizer.hasMoreTokens() &&
-      (!(this.tokenizer.tokenType() === "symbol") ||
-        !(this.tokenizer.symbol() === ";"))
+      this.tokenizer.currentToken() !== ";"
     ) {
       this.convertToken();
     }
+
     this.endBlock(tag);
   }
 
@@ -149,37 +138,31 @@ export class CompilationEngine {
    * (’constructor’ | ’function’ | ’method’) (’void’ | type) subroutineName ’(’ parameterList ’)’ subroutineBody
    */
   compileSubroutine(): void {
-    const tag = "subroutineDec";
-    this.startBlock(tag);
+    const decTag = "subroutineDec";
+    this.startBlock(decTag);
+    const bodyTag = "subroutineBody";
 
     const type = this.tokenizer.tokenType();
-    // <keyword> { constructor | function | method } </keyword>
+    // <keyword> constructor/function/method </keyword>
     this.pushResults(`<${type}> ${this.tokenizer.keyWord()} </${type}>`);
     while (
       this.tokenizer.hasMoreTokens() &&
-      (!(this.tokenizer.tokenType() === "symbol") ||
-        !(this.tokenizer.symbol() === "}"))
+      this.tokenizer.nextToken() !== "}"
     ) {
-      if (
-        this.tokenizer.tokenType() === "symbol" &&
-        this.tokenizer.symbol() === "("
-      ) {
+      if (this.tokenizer.currentToken() === "(") {
         this.compileParameterList();
+        this.convertToken(); // ")" を出力
         continue;
-      } else if (
-        this.tokenizer.tokenType() === "symbol" &&
-        this.tokenizer.symbol() === "{"
-      ) {
-        // { は subroutineBody の要素に入れる
-        const last = this.popResults();
-        this.startBlock("subroutineBody");
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        this.pushResults(last!);
+      } else if (this.tokenizer.currentToken() === ")") {
+        this.startBlock(bodyTag);
+        this.convertToken(); // "{" を出力
+        continue;
       }
       this.convertToken();
     }
-    this.endBlock("subroutineBody");
-    this.endBlock(tag);
+    this.convertToken(); // "}" を出力
+    this.endBlock(bodyTag);
+    this.endBlock(decTag);
   }
 
   /**
@@ -193,18 +176,12 @@ export class CompilationEngine {
 
     while (
       this.tokenizer.hasMoreTokens() &&
-      !(
-        this.tokenizer.tokenType() === "symbol" &&
-        this.tokenizer.symbol() === ")"
-      )
+      this.tokenizer.nextToken() !== ")"
     ) {
       this.convertToken();
     }
-    // ) は parameterList の要素に入れない
-    const last = this.popResults();
+
     this.endBlock(tag);
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    this.pushResults(last!);
   }
 
   /**
@@ -221,11 +198,11 @@ export class CompilationEngine {
     this.pushResults(`<${type}> ${this.tokenizer.keyWord()} </${type}>`);
     while (
       this.tokenizer.hasMoreTokens() &&
-      (!(this.tokenizer.tokenType() === "symbol") ||
-        !(this.tokenizer.symbol() === ";"))
+      this.tokenizer.currentToken() !== ";"
     ) {
       this.convertToken();
     }
+
     this.endBlock(tag);
   }
 
@@ -240,17 +217,14 @@ export class CompilationEngine {
     const tag = "statements";
     this.startBlock(tag);
 
-    // FIXME: JackTokenizer.keyWord() 呼び出しのため、適切なトークンまで進める
-    if (
-      this.tokenizer.hasMoreTokens() &&
-      this.tokenizer.tokenType() !== "keyword"
-    ) {
+    // FIXME? 適切な場所までトークンを進める
+    if (this.tokenizer.tokenType() !== "keyword") {
       this.tokenizer.advance();
     }
 
-    let currentToken = this.tokenizer.keyWord();
-    while (["do", "let", "while", "return", "if"].includes(currentToken)) {
-      switch (currentToken) {
+    let token = this.tokenizer.currentToken();
+    while (["do", "let", "while", "return", "if"].includes(token)) {
+      switch (token) {
         case "do":
           this.compileDo();
           break;
@@ -267,7 +241,7 @@ export class CompilationEngine {
           this.compileIf();
           break;
       }
-      currentToken = this.tokenizer.currentToken();
+      token = this.tokenizer.nextToken();
     }
     this.endBlock(tag);
   }
@@ -281,30 +255,27 @@ export class CompilationEngine {
     const tag = "doStatement";
     this.startBlock(tag);
 
-    // FIXME: JackTokenizer.keyWord() 呼び出しのため、適切なトークンまで進める
-    while (
-      this.tokenizer.hasMoreTokens() &&
-      this.tokenizer.tokenType() !== "keyword"
-    ) {
+    // statement が2つ以上連続するときに、";"からトークンを進める
+    if (this.tokenizer.tokenType() !== "keyword") {
       this.tokenizer.advance();
     }
 
     const type = this.tokenizer.tokenType();
     // <keyword> do </keyword>
     this.pushResults(`<${type}> ${this.tokenizer.keyWord()} </${type}>`);
+
     while (
       this.tokenizer.hasMoreTokens() &&
-      (!(this.tokenizer.tokenType() === "symbol") ||
-        !(this.tokenizer.symbol() === ";"))
+      this.tokenizer.currentToken() !== ";"
     ) {
-      if (
-        this.tokenizer.tokenType() === "symbol" &&
-        this.tokenizer.symbol() === "("
-      ) {
+      if (this.tokenizer.currentToken() === "(") {
         this.compileExpressionList();
+        this.convertToken(); // ")" を出力
+        continue;
       }
       this.convertToken();
     }
+
     this.endBlock(tag);
   }
 
@@ -317,35 +288,29 @@ export class CompilationEngine {
     const tag = "letStatement";
     this.startBlock(tag);
 
-    // FIXME: JackTokenizer.keyWord() 呼び出しのため、適切なトークンまで進める
-    if (
-      this.tokenizer.hasMoreTokens() &&
-      this.tokenizer.tokenType() !== "keyword"
-    ) {
+    // statement が2つ以上連続するときに、";"からトークンを進める
+    if (this.tokenizer.tokenType() !== "keyword") {
       this.tokenizer.advance();
     }
 
     const type = this.tokenizer.tokenType();
     // <keyword> let </keyword>
     this.pushResults(`<${type}> ${this.tokenizer.keyWord()} </${type}>`);
+
     while (
       this.tokenizer.hasMoreTokens() &&
-      (!(this.tokenizer.tokenType() === "symbol") ||
-        !(this.tokenizer.symbol() === ";"))
+      this.tokenizer.currentToken() !== ";"
     ) {
       if (
-        this.tokenizer.tokenType() === "symbol" &&
-        this.tokenizer.symbol() === "["
+        this.tokenizer.currentToken() === "[" ||
+        this.tokenizer.currentToken() === "="
       ) {
         this.compileExpression();
-      } else if (
-        this.tokenizer.tokenType() === "symbol" &&
-        this.tokenizer.symbol() === "="
-      ) {
-        this.compileExpression();
+        continue;
       }
       this.convertToken();
     }
+
     this.endBlock(tag);
   }
 
@@ -358,11 +323,8 @@ export class CompilationEngine {
     const tag = "whileStatement";
     this.startBlock(tag);
 
-    // FIXME: JackTokenizer.keyWord() 呼び出しのため、適切なトークンまで進める
-    while (
-      this.tokenizer.hasMoreTokens() &&
-      this.tokenizer.tokenType() !== "keyword"
-    ) {
+    // statement が2つ以上連続するときに、";"からトークンを進める
+    if (this.tokenizer.tokenType() !== "keyword") {
       this.tokenizer.advance();
     }
 
@@ -372,30 +334,21 @@ export class CompilationEngine {
 
     while (
       this.tokenizer.hasMoreTokens() &&
-      (this.tokenizer.tokenType() !== "symbol" ||
-        this.tokenizer.symbol() !== "}")
+      this.tokenizer.currentToken() !== "}"
     ) {
-      this.convertToken();
-      if (
-        this.tokenizer.tokenType() === "symbol" &&
-        this.tokenizer.symbol() === "("
-      ) {
+      if (this.tokenizer.currentToken() === "(") {
         this.compileExpression();
-      } else if (
-        this.tokenizer.tokenType() === "symbol" &&
-        this.tokenizer.symbol() === "{"
-      ) {
-        // {} 内の statements がなければ {} の間には何も入れない
-        if (this.tokenizer.currentToken() === "}") {
-          this.convertToken();
-          this.endBlock(tag);
-          return;
-        }
+        this.convertToken(); // ")" を出力
+        continue;
+      } else if (this.tokenizer.currentToken() === ")") {
+        this.convertToken(); // "{" を出力
         this.compileStatements();
-        // <symbol> } </symbol>
-        this.convertToken();
+        this.convertToken(); // "}" を出力
+        continue;
       }
+      this.convertToken();
     }
+
     this.endBlock(tag);
   }
 
@@ -408,11 +361,8 @@ export class CompilationEngine {
     const tag = "returnStatement";
     this.startBlock(tag);
 
-    // FIXME: JackTokenizer.keyWord() 呼び出しのため、適切なトークンまで進める
-    while (
-      this.tokenizer.hasMoreTokens() &&
-      this.tokenizer.tokenType() !== "keyword"
-    ) {
+    // statement が2つ以上連続するときに、";"からトークンを進める
+    if (this.tokenizer.tokenType() !== "keyword") {
       this.tokenizer.advance();
     }
 
@@ -420,13 +370,14 @@ export class CompilationEngine {
     // <keyword> return </keyword>
     this.pushResults(`<${type}> ${this.tokenizer.keyWord()} </${type}>`);
 
-    let currentToken = this.tokenizer.currentToken();
-    while (currentToken !== ";") {
+    while (
+      this.tokenizer.hasMoreTokens() &&
+      this.tokenizer.nextToken() !== ";"
+    ) {
       this.compileExpression();
-      currentToken = this.tokenizer.currentToken();
     }
-    // <symbol> ; </symbol>
-    this.convertToken();
+    this.convertToken(); // ";"を出力
+
     this.endBlock(tag);
   }
 
@@ -439,11 +390,8 @@ export class CompilationEngine {
     const tag = "ifStatement";
     this.startBlock(tag);
 
-    // FIXME: JackTokenizer.keyWord() 呼び出しのため、適切なトークンまで進める
-    while (
-      this.tokenizer.hasMoreTokens() &&
-      this.tokenizer.tokenType() !== "keyword"
-    ) {
+    // statement が2つ以上連続するときに、";"からトークンを進める
+    if (this.tokenizer.tokenType() !== "keyword") {
       this.tokenizer.advance();
     }
 
@@ -451,40 +399,23 @@ export class CompilationEngine {
     // <keyword> if </keyword>
     this.pushResults(`<${type}> ${this.tokenizer.keyWord()} </${type}>`);
 
-    let currentToken = this.tokenizer.currentToken();
     while (
       this.tokenizer.hasMoreTokens() &&
-      !(
-        this.tokenizer.tokenType() === "symbol" &&
-        this.tokenizer.symbol() === "}" &&
-        // } の後に else が続く場合は、if 文のコンパイルを続ける
-        currentToken !== "else"
-      )
+      (this.tokenizer.currentToken() !== "}" ||
+        (this.tokenizer.currentToken() === "}" &&
+          this.tokenizer.nextToken() === "else"))
     ) {
-      if (
-        this.tokenizer.tokenType() === "symbol" &&
-        this.tokenizer.symbol() === "("
-      ) {
+      if (this.tokenizer.currentToken() === "(") {
         this.compileExpression();
-      } else if (
-        this.tokenizer.tokenType() === "symbol" &&
-        this.tokenizer.symbol() === "{"
-      ) {
-        if (this.tokenizer.currentToken() === "}") {
-          // {} 内の statements がなくても、statements のタグは入れる
-          const t = "statements";
-          this.startBlock(t);
-          this.endBlock(t);
-          // <symbol> } </symbol>
-          this.convertToken();
-        } else {
-          this.compileStatements();
-          // this.tokenizer.advance();
-        }
-      } else {
-        this.convertToken();
+        this.convertToken(); // ")" を出力
+        continue;
+      } else if (this.tokenizer.currentToken() === ")") {
+        this.convertToken(); // "{" を出力
+        this.compileStatements();
+        this.convertToken(); // "}" を出力
+        continue;
       }
-      currentToken = this.tokenizer.currentToken();
+      this.convertToken();
     }
 
     this.endBlock(tag);
@@ -499,26 +430,14 @@ export class CompilationEngine {
     const tag = "expression";
     this.startBlock(tag);
 
-    let count = 0;
-    let currentToken = this.tokenizer.currentToken();
+    this.compileTerm();
     while (
       this.tokenizer.hasMoreTokens() &&
-      ![")", "]", ";", ","].includes(currentToken)
+      operators.includes(this.tokenizer.nextToken())
     ) {
-      const type = this.tokenizer.tokenType();
-      if (type === "keyword" || type === "symbol") {
-        // expression の途中に出てくる演算子は、term に含まない
-        if (count > 0 && operators.includes(currentToken)) {
-          this.convertToken();
-        }
-        this.compileTerm();
-      } else {
-        this.convertToken();
-      }
-      currentToken = this.tokenizer.currentToken();
-      count++;
+      this.convertToken();
+      this.compileTerm();
     }
-
     this.endBlock(tag);
   }
 
@@ -531,53 +450,69 @@ export class CompilationEngine {
     const tag = "term";
     this.startBlock(tag);
 
-    let currentToken = this.tokenizer.currentToken();
-    // かっこ始まりの並びのとき、最後にかっこで閉じるための準備
-    let endMark;
-    switch (currentToken) {
+    let endMarks;
+    switch (this.tokenizer.nextToken()) {
       case "(":
-        endMark = [")"];
+        endMarks = [")"];
         break;
       case "[":
-        endMark = ["]"];
+        endMarks = ["]"];
         break;
       default:
-        endMark = [")", "]", ";", ","];
+        endMarks = [")", "]", ";", ","];
     }
-    while (this.tokenizer.hasMoreTokens() && !endMark.includes(currentToken)) {
+
+    while (
+      this.tokenizer.hasMoreTokens() &&
+      !endMarks.includes(this.tokenizer.nextToken())
+    ) {
       if (this.tokenizer.tokenType() === "identifier") {
-        switch (this.tokenizer.currentToken()) {
+        switch (this.tokenizer.nextToken()) {
           case "[":
-            // <symbol> [ </symbol>
-            this.convertToken();
+            // 配列 varName ’[’ expression ’]’
+            this.convertToken(); // "["を出力
             this.compileExpression();
+            this.convertToken(); // "]"を出力
             break;
           case "(":
-            // <symbol> ( </symbol>
-            this.convertToken();
+            // サブルーチン呼び出しの引数
+            this.convertToken(); // "("を出力
             this.compileExpressionList();
+            this.convertToken(); // ")"を出力
             break;
           case ".":
+            // クラス、オブジェクトのサブルーチン呼び出し
+            this.convertToken(); // "."を出力
+            this.convertToken(); // クラス/オブジェクト名を出力
+            this.convertToken(); // "("を出力
+            this.compileExpressionList();
+            this.convertToken(); // ")"を出力
             break;
           default:
-            // 配列、サブルーチン呼び出しではないとき
-            this.endBlock(tag);
-            return;
+          // 変数のときは何もしない
         }
-      } else if (unaryOperators.includes(currentToken)) {
-        this.convertToken();
+        this.endBlock(tag);
+        return;
+      }
+      // unaryOp + term
+      if (unaryOperators.includes(this.tokenizer.nextToken())) {
+        this.convertToken(); // unaryOp
         this.compileTerm();
         this.endBlock(tag);
         return;
-      } else if (["(", "["].includes(currentToken)) {
-        this.convertToken();
-        this.compileExpression();
-        this.convertToken();
-        break;
       }
+      // ’(’ expression ’)’
+      if (this.tokenizer.nextToken() === "(") {
+        this.convertToken(); // "("を出力
+        this.compileExpression();
+        this.convertToken(); // ")"を出力
+        this.endBlock(tag);
+        return;
+      }
+
       this.convertToken();
-      currentToken = this.tokenizer.currentToken();
     }
+
     this.endBlock(tag);
   }
 
@@ -590,16 +525,17 @@ export class CompilationEngine {
     const tag = "expressionList";
     this.startBlock(tag);
 
-    let currentToken = this.tokenizer.currentToken();
-    while (this.tokenizer.hasMoreTokens() && ![")"].includes(currentToken)) {
-      if (currentToken === ",") {
-        // <symbol> , </symbol>
+    while (
+      this.tokenizer.hasMoreTokens() &&
+      this.tokenizer.nextToken() !== ")"
+    ) {
+      if (this.tokenizer.nextToken() === ",") {
         this.convertToken();
-      } else {
-        this.compileExpression();
+        continue;
       }
-      currentToken = this.tokenizer.currentToken();
+      this.compileExpression();
     }
+
     this.endBlock(tag);
   }
 }
